feat(jf-playback-reporting): add getMediaServerAdmin util function
